Refactor app to module, add plugins via composition
