+++ zabbix_cli/app/__init__.py
@@ -1,0 +1,3 @@
+from __future__ import annotations
+
+from .app import *  # noqa: F403 # refactor from file to module
